Moved FlipIn/OutXTransition to controls and used that implementation thereby reducing code duplications of these two transormations fixes TweetWallFX/TweetwallFX#401
