+++ devoxx-2018-pl/cinema/src/main/java/org/tweetwallfx/devoxx18pl/steps/DevoxxShowTopRatedToday.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  *
- * Copyright 2017 TweetWallFX
+ * Copyright 2017-2018 TweetWallFX
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -45,7 +45,7 @@ import org.tweetwallfx.controls.dataprovider.DataProvider;
 import org.tweetwallfx.controls.stepengine.Step;
 import org.tweetwallfx.controls.stepengine.StepEngine.MachineContext;
 import org.tweetwallfx.controls.stepengine.config.StepEngineSettings;
-import org.tweetwallfx.devoxx18pl.animations.FlipInXTransition;
+import org.tweetwallfx.controls.transition.FlipInXTransition;
 import org.tweetwallfx.devoxx18pl.dataprovider.SpeakerImageProvider;
 import org.tweetwallfx.devoxx18pl.dataprovider.TopTalksTodayDataProvider;
 import org.tweetwallfx.devoxx18pl.dataprovider.VotedTalk;
