+++ voxxed-days-zurich/src/main/java/org/tweetwallfx/vdz/steps/FlipOutScheduleToday.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  *
- * Copyright 2017 TweetWallFX
+ * Copyright 2017-2018 TweetWallFX
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -23,13 +23,12 @@
  */
 package org.tweetwallfx.vdz.steps;
 
+import javafx.scene.Node;
 import org.tweetwallfx.controls.WordleSkin;
 import org.tweetwallfx.controls.stepengine.Step;
 import org.tweetwallfx.controls.stepengine.StepEngine.MachineContext;
 import org.tweetwallfx.controls.stepengine.config.StepEngineSettings;
-import org.tweetwallfx.vdz.animation.FlipOutXTransition;
-
-import javafx.scene.Node;
+import org.tweetwallfx.controls.transition.FlipOutXTransition;
 
 /**
  * Schedule Flip Out Animation Step
